Image draws in chunks to allow large images
This is gonna be the last update. I tried to make image processing as efficient as possible while taking into account the possibility of the user importing large images. The program can still crash if the image is really large but I think for a Flash program it does the job very very well.